pyproject.toml
[project]
name = "dragonfly-mainframe"
description = "Distributes tasks to dragonfly-clients and handles package status updates."
authors = [
    { name = "Vipyr Security Developers", email = "support@vipyrsec.com" },
]
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.95.1",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.22.0",
    "SQLAlchemy>=2.0.12",
    "python-dotenv>=1.0.0",
    "alembic>=1.10.4",
    "letsbuilda-pypi==5.0.0b3",
    "msgraph-core>=0.2.2",
    "azure-identity>=1.13.0",
    "structlog>=23.1.0",
    "asgi-correlation-id>=4.2.0",
    "sentry-sdk[fastapi]>=1.25.1",
    "structlog-sentry>=2.0.3",
    "requests>=2.30.0",
    "httpx>=0.24.1",
    "psycopg2-binary",
]
dynamic = ["version"]

[project.urls]
repository = "https://github.com/vipyrsec/dragonfly-mainframe/"
documentation = "https://docs.vipyrsec.dev/dragonfly-mainframe/"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
version = { source = "file", path = "src/mainframe/__init__.py" }

[tool.pdm.dev-dependencies]
lint = [
    "pre-commit>=3.3.1",
    "ruff>=0.0.265",
    "black>=23.3.0",
    "isort>=5.12.0",
    "pyright>=1.1.306",
]
test = [
    "pytest>=7.3.1",
    "coverage>=7.2.7",
]
logs = ["rich>=13.4.2"]
docs = [
    "sphinx",
    "furo",
    "sphinx-autoapi",
    "releases",
]

[tool.pdm.scripts]
format = { shell = "black . && isort ." }
lint = { shell = "pyright; ruff ." }
test = { shell = "docker compose -f compose.yaml -f compose-tests.yaml up --build --exit-code-from mainframe" }
coverage = { shell = "coverage run -m pytest tests -vv && coverage report && coverage xml" }
precommit = { shell = "pre-commit run --all-files" }
all = { composite = ["format", "lint", "test"] }

start = { cmd = "docker compose up --build" }

[tool.isort]
profile = "black"

[tool.black]
target-version = ["py311"]
line-length = 120

[tool.ruff]
line-length = 120

[tool.pyright]
include = ["src/**/*.py"]
ignore = ["alembic/"]

venvPath = "."
venv = ".venv"


strictListInference = true
strictDictionaryInference = true
strictSetInference = true
analyzeUnannotatedFunctions = true
strictParameterNoneValue = true
enableTypeIgnoreComments = true
reportMissingModuleSource = "warning"
reportMissingImports = "error"
reportUndefinedVariable = "error"
reportAssertAlwaysTrue = "error"
reportInvalidStringEscapeSequence = "error"
reportInvalidTypeVarUse = "error"
reportMissingTypeStubs = "none"
reportSelfClsParameterName = "error"
reportUnsupportedDunderAll = "error"
reportUnusedExpression = "error"
reportWildcardImportFromLibrary = "error"
reportGeneralTypeIssues = "error"
reportOptionalSubscript = "error"
reportOptionalMemberAccess = "error"
reportOptionalCall = "error"
reportOptionalIterable = "error"
reportOptionalContextManager = "error"
reportOptionalOperand = "error"
reportTypedDictNotRequiredAccess = "error"
reportPrivateImportUsage = "error"
reportUnboundVariable = "error"
reportUnusedCoroutine = "error"
reportConstantRedefinition = "error"
reportDeprecated = "error"
reportDuplicateImport = "error"
reportFunctionMemberAccess = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportIncompleteStub = "error"
reportInconsistentConstructor = "error"
reportInvalidStubStatement = "error"
reportMatchNotExhaustive = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = "error"
reportOverlappingOverload = "error"
reportPrivateUsage = "error"
reportTypeCommentUsage = "error"
reportUnknownArgumentType = "error"
reportUnknownLambdaType = "error"
reportUnknownMemberType = "error"
reportUnknownParameterType = "error"
reportUnknownVariableType = "error"
reportUnnecessaryCast = "error"
reportUnnecessaryComparison = "error"
reportUnnecessaryContains = "error"
reportUnnecessaryIsInstance = "error"
reportUnusedClass = "error"
reportUnusedImport = "error"
reportUnusedFunction = "error"
reportUnusedVariable = "error"
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"

[tool.coverage.run]
branch = false
omit = [
    "__pypackages__/*",
    "src/mainframe/dependencies.py",
    "src/mainframe/database.py",
    "src/mainframe/json_web_token.py",
    "src/mainframe/custom_exceptions.py",
    "src/mainframe/utils/microsoft.py",
    "src/mainframe/utils/mailer.py"
]

[tool.coverage.report]
fail_under = 100
